crawling_scripts/requirements.txt
IMDbPY==2022.7.9
IMDbPY==2022.7.9
matplotlib==3.7.1
pandas==2.0.0
ratelimit==2.2.1
seaborn==0.12.2
